fix: bullet c++/wasm mostly fixes

+++ emscripten/bullet/src/demo/states/IState.hpp
@@ -14,7 +14,7 @@ public:
 
 public:
   virtual void handleEvent(const SDL_Event&) = 0;
-  virtual void update(int) = 0;
+  virtual void update(float) = 0;
   virtual void render(const SDL_Window&) = 0;
   virtual void resize(int, int) = 0;
   virtual void visibility(bool) = 0;
