+++ emscripten/bullet/src/demo/states/State_Paused.hpp
@@ -6,6 +6,6 @@
 class State_Paused : public State_AbstractSimulation {
 public:
   virtual void handleEvent(const SDL_Event&) override;
-  virtual void update(int) override;
+  virtual void update(float) override;
   virtual void visibility(bool visible) override;
 };
