+++ emscripten/bullet/src/demo/states/State_Running.hpp
@@ -10,5 +10,5 @@ public:
   virtual void leave() override;
 
 public:
-  virtual void update(int) override;
+  virtual void update(float) override;
 };
