+++ emscripten/bullet/src/demo/states/State_StartGeneration.hpp
@@ -14,5 +14,5 @@ public:
   virtual void leave() override;
 
 public:
-  virtual void update(int) override;
+  virtual void update(float) override;
 };
