+++ emscripten/bullet/src/demo/states/State_WebWorkersLoading.hpp
@@ -19,7 +19,7 @@ public:
 
 public:
   virtual void handleEvent(const SDL_Event&) override;
-  virtual void update(int) override;
+  virtual void update(float) override;
   virtual void render(const SDL_Window&) override;
 
   virtual void visibility(bool visible) override;
